requirements.txt
cachelib==0.9.0
click==8.1.3
colorama==0.4.5
cs50==9.2.2
DateTime==4.7
Flask==2.2.2
flask-session2==1.3.1
greenlet==1.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
passlib==1.7.4
pytz==2022.2.1
SQLAlchemy==1.4.41
sqlparse==0.4.2
termcolor==1.1.0
Werkzeug==2.2.2
zope.interface==5.4.0